apps/cli: switch agent to permissionMode 'auto' and drop ad-hoc security gating
The Claude Agent SDK 0.2.117 added an 'auto' permission mode that uses a
model classifier to approve/deny tool calls. Switch to it and remove the
in-house path-gated approval flow (TRUSTED_ROOTS, approvedPermissions in
cli.json, ALLOWED_TOOLS list, autoApprove flag) — the classifier
supersedes them.

AskUserQuestion still needs to render in our chat UI rather than the
SDK's default prompt, so route it through a PreToolUse hook scoped with
matcher: 'AskUserQuestion' that injects answers via updatedInput. Every
other tool falls through to the classifier.
