+++ apps/cli/commands/ai/tests/ai.test.ts
@@ -729,11 +729,6 @@ describe( 'CLI: studio code --json mode', () => {
 			type: 'turn.completed',
 			status: 'success',
 		} );
-		expect( startAiAgent ).toHaveBeenCalledWith(
-			expect.objectContaining( {
-				autoApprove: true,
-			} )
-		);
 		expect( process.exitCode ).not.toBe( 1 );
 	} );
 
@@ -807,50 +802,5 @@ describe( 'CLI: studio code --json mode', () => {
 			status: 'error',
 		} );
 		expect( process.exitCode ).toBe( 1 );
-	} );
-
-	it( 'does not call autoApprove in interactive mode', async () => {
-		waitForInputMock.mockResolvedValueOnce( 'Hello' ).mockResolvedValueOnce( '/exit' );
-
-		await buildParser().parseAsync( [ 'ai' ] );
-
-		expect( startAiAgent ).toHaveBeenCalledWith(
-			expect.objectContaining( {
-				autoApprove: false,
-			} )
-		);
-	} );
-
-	it( 'respects --no-auto-approve flag independently of --json', async () => {
-		const resultMessage = {
-			type: 'result' as const,
-			subtype: 'success' as const,
-			session_id: 'auto-approve-test',
-			num_turns: 1,
-			total_cost_usd: 0.001,
-		};
-		vi.mocked( startAiAgent ).mockReturnValueOnce( {
-			interrupt: vi.fn().mockResolvedValue( undefined ),
-			[ Symbol.asyncIterator ]() {
-				let emitted = false;
-				return {
-					next: async () => {
-						if ( ! emitted ) {
-							emitted = true;
-							return { done: false as const, value: resultMessage };
-						}
-						return { done: true as const, value: undefined };
-					},
-				};
-			},
-		} as never );
-
-		await buildParser().parseAsync( [ 'ai', 'hello', '--json', '--no-auto-approve' ] );
-
-		expect( startAiAgent ).toHaveBeenCalledWith(
-			expect.objectContaining( {
-				autoApprove: false,
-			} )
-		);
 	} );
 } );
